move multiqueue metrics reporter into own package to prevent circular deps
