+++ enterprise/cmd/worker/internal/executormultiqueue/BUILD.bazel
@@ -1,0 +1,19 @@
+load("@io_bazel_rules_go//go:def.bzl", "go_library")
+
+go_library(
+    name = "executormultiqueue",
+    srcs = ["multiqueue_metrics_reporter.go"],
+    importpath = "github.com/sourcegraph/sourcegraph/enterprise/cmd/worker/internal/executormultiqueue",
+    visibility = ["//enterprise/cmd/worker:__subpackages__"],
+    deps = [
+        "//cmd/worker/job",
+        "//cmd/worker/shared/init/db",
+        "//enterprise/cmd/worker/internal/batches",
+        "//enterprise/cmd/worker/internal/executorqueue",
+        "//enterprise/internal/codeintel/autoindexing",
+        "//internal/env",
+        "//internal/goroutine",
+        "//internal/observation",
+        "//internal/workerutil/dbworker/store",
+    ],
+)
