+++ enterprise/cmd/worker/internal/executorqueue/BUILD.bazel
@@ -8,7 +8,6 @@ go_library(
         "config.go",
         "external_emitter.go",
         "gcp_reporter.go",
-        "multiqueue_metrics_reporter.go",
         "prometheus.go",
         "queue_allocation.go",
         "reporter.go",
@@ -16,10 +15,6 @@ go_library(
     importpath = "github.com/sourcegraph/sourcegraph/enterprise/cmd/worker/internal/executorqueue",
     visibility = ["//enterprise/cmd/worker:__subpackages__"],
     deps = [
-        "//cmd/worker/job",
-        "//cmd/worker/shared/init/db",
-        "//enterprise/cmd/worker/internal/batches",
-        "//enterprise/internal/codeintel/autoindexing",
         "//enterprise/internal/codeintel/uploads/shared",
         "//enterprise/internal/executor/util",
         "//internal/env",
