+++ enterprise/cmd/worker/shared/shared.go
@@ -16,6 +16,7 @@ import (
 	"github.com/sourcegraph/sourcegraph/enterprise/cmd/worker/internal/codeintel"
 	"github.com/sourcegraph/sourcegraph/enterprise/cmd/worker/internal/codemonitors"
 	repoembeddings "github.com/sourcegraph/sourcegraph/enterprise/cmd/worker/internal/embeddings/repo"
+	"github.com/sourcegraph/sourcegraph/enterprise/cmd/worker/internal/executormultiqueue"
 	"github.com/sourcegraph/sourcegraph/enterprise/cmd/worker/internal/executors"
 	workerinsights "github.com/sourcegraph/sourcegraph/enterprise/cmd/worker/internal/insights"
 	"github.com/sourcegraph/sourcegraph/enterprise/cmd/worker/internal/permissions"
@@ -31,22 +32,23 @@ import (
 )
 
 var additionalJobs = map[string]job.Job{
-	"codehost-version-syncing":      versions.NewSyncingJob(),
-	"insights-job":                  workerinsights.NewInsightsJob(),
-	"insights-query-runner-job":     workerinsights.NewInsightsQueryRunnerJob(),
-	"insights-data-retention-job":   workerinsights.NewInsightsDataRetentionJob(),
-	"batches-janitor":               batches.NewJanitorJob(),
-	"batches-scheduler":             batches.NewSchedulerJob(),
-	"batches-reconciler":            batches.NewReconcilerJob(),
-	"batches-bulk-processor":        batches.NewBulkOperationProcessorJob(),
-	"batches-workspace-resolver":    batches.NewWorkspaceResolverJob(),
-	"executors-janitor":             executors.NewJanitorJob(),
-	"executors-metricsserver":       executors.NewMetricsServerJob(),
-	"codemonitors-job":              codemonitors.NewCodeMonitorJob(),
-	"bitbucket-project-permissions": permissions.NewBitbucketProjectPermissionsJob(),
-	"permission-sync-job-cleaner":   permissions.NewPermissionSyncJobCleaner(),
-	"permission-sync-job-scheduler": permissions.NewPermissionSyncJobScheduler(),
-	"export-usage-telemetry":        telemetry.NewTelemetryJob(),
+	"codehost-version-syncing":              versions.NewSyncingJob(),
+	"insights-job":                          workerinsights.NewInsightsJob(),
+	"insights-query-runner-job":             workerinsights.NewInsightsQueryRunnerJob(),
+	"insights-data-retention-job":           workerinsights.NewInsightsDataRetentionJob(),
+	"batches-janitor":                       batches.NewJanitorJob(),
+	"batches-scheduler":                     batches.NewSchedulerJob(),
+	"batches-reconciler":                    batches.NewReconcilerJob(),
+	"batches-bulk-processor":                batches.NewBulkOperationProcessorJob(),
+	"batches-workspace-resolver":            batches.NewWorkspaceResolverJob(),
+	"executors-janitor":                     executors.NewJanitorJob(),
+	"executors-metricsserver":               executors.NewMetricsServerJob(),
+	"executors-multiqueue-metrics-reporter": executormultiqueue.NewMultiqueueMetricsReporterJob(),
+	"codemonitors-job":                      codemonitors.NewCodeMonitorJob(),
+	"bitbucket-project-permissions":         permissions.NewBitbucketProjectPermissionsJob(),
+	"permission-sync-job-cleaner":           permissions.NewPermissionSyncJobCleaner(),
+	"permission-sync-job-scheduler":         permissions.NewPermissionSyncJobScheduler(),
+	"export-usage-telemetry":                telemetry.NewTelemetryJob(),
 
 	"codeintel-policies-repository-matcher":       codeintel.NewPoliciesRepositoryMatcherJob(),
 	"codeintel-autoindexing-summary-builder":      codeintel.NewAutoindexingSummaryBuilder(),
